// mono-repo/packages/gantry-app/src/app/boxing/models/boxing-constants.model.ts
export enum BoxingEnumContent {
    AdditionalInfo = 'AdditionalInfo',
    MoreMarkets = 'MoreMarkets',
    OnRequest = 'OnRequest',
    OptionalInfo = 'OptionalInfo',
    LeadTitle = 'LeadTitle',
}

export enum BoxingMarkets {
    FIGHTBETTING = 'FightBetting',
    ROUNDBETTING = 'RoundBetting',
    ROUNDGROUPBETTING = 'RoundGroupBetting',
    METHODOFVICTORY = 'MethodOfVictory',
}

export enum SelectionName {
    DRAWORTECHNICALDRAW = 'DRAW OR TECHNICAL DRAW',
    DRAW = 'DRAW',
    POINTSNAME = 'POINTS (FULL DISTANCE)',
}

export enum AdvantageType {
    Home = 'H',
    Away = 'A',
}

export enum Result {
    Draw = 'D',
}